fix: unable to delete user account. Issue #402 (#413)
* fix: unable to delete user account. Issue #402

* chore: fix checkstyle and import orders

* ci: fix sonar project key

* chore: improve test condition and test cases

+++ server/src/main/java/br/com/tasknoteapp/server/service/UserSessionService.java
@@ -1,14 +1,22 @@
 package br.com.tasknoteapp.server.service;
 
+import br.com.tasknoteapp.server.entity.NotesCreatedEntity;
+import br.com.tasknoteapp.server.entity.UserEntity;
+import br.com.tasknoteapp.server.entity.UserTasksDone;
+import br.com.tasknoteapp.server.exception.UserNotFoundException;
+import br.com.tasknoteapp.server.repository.NotesCreatedRepository;
+import br.com.tasknoteapp.server.repository.UserTasksDoneRepository;
 import br.com.tasknoteapp.server.response.JwtAuthenticationResponse;
 import br.com.tasknoteapp.server.response.NoteResponse;
 import br.com.tasknoteapp.server.response.TaskResponse;
 import br.com.tasknoteapp.server.response.UserResponse;
 import jakarta.transaction.Transactional;
 import java.util.List;
+import java.util.Optional;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+/** This class contains methods to handle user session and account deletion. */
 @Service
 @AllArgsConstructor
 public class UserSessionService {
@@ -18,6 +26,10 @@ public class UserSessionService {
   private final TaskService taskService;
 
   private final NoteService noteService;
+
+  private final UserTasksDoneRepository userTasksDoneRepository;
+
+  private final NotesCreatedRepository notesCreatedRepository;
 
   /**
    * Refresh the current user session with a new JWT token.
@@ -36,6 +48,11 @@ public class UserSessionService {
    */
   @Transactional
   public UserResponse deleteCurrentUserAccount() {
+    Optional<UserEntity> userOptional = authService.getCurrentUser();
+    if (userOptional.isEmpty()) {
+      throw new UserNotFoundException();
+    }
+
     List<TaskResponse> tasks = taskService.getAllTasks();
     for (TaskResponse task : tasks) {
       taskService.deleteTask(task.id());
@@ -44,6 +61,17 @@ public class UserSessionService {
     List<NoteResponse> notes = noteService.getAllNotes();
     for (NoteResponse note : notes) {
       noteService.deleteNote(note.id());
+    }
+
+    Long userId = userOptional.get().getId();
+    List<UserTasksDone> doneTasks = userTasksDoneRepository.findAllById_userId(userId);
+    if (!doneTasks.isEmpty()) {
+      userTasksDoneRepository.deleteAllInBatch(doneTasks);
+    }
+
+    List<NotesCreatedEntity> notesStats = notesCreatedRepository.findAllByUserId(userId);
+    if (!notesStats.isEmpty()) {
+      notesCreatedRepository.deleteAll(notesStats);
     }
 
     return authService.deleteUserAccount();
